Optimization of code
- 优化work_1_correct( )批阅循环条件
- 重构driver_init.py
- 优化main.py
